# apps/app/swagger-spec.json
{"openapi":"3.0.0","paths":{"/reports/initiate":{"get":{"description":"Initiate fetching reports for target interval","operationId":"ReportsController_initiateReports","parameters":[{"name":"from_date","required":true,"in":"query","description":"The start date of the reports to fetch","schema":{"example":"2024-11-25T00:00:00.000Z","type":"string"}},{"name":"to_date","required":true,"in":"query","description":"The end date of the reports to fetch","schema":{"example":"2024-11-26T00:00:00.000Z","type":"string"}},{"name":"event_name","required":true,"in":"query","description":"The event name of the reports to fetch","schema":{"example":"install","enum":["install","purchase"],"type":"string"}}],"responses":{"200":{"description":""},"400":{"description":"Validation error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpErrorSchema"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpErrorSchema"},"example":{"message":"Internal server error","statusCode":500}}}}},"summary":"Initiate fetching reports","tags":["Reports"]}},"/reports/aggregated/ad_id":{"get":{"description":"Get aggregated reports by ad_id","operationId":"ReportsController_getAggregatedReportsByAdId","parameters":[{"name":"from_date","required":true,"in":"query","description":"The start date of the reports","schema":{"example":"2024-11-25T00:00:00.000Z","type":"string"}},{"name":"to_date","required":true,"in":"query","description":"The end date of the reports","schema":{"example":"2024-11-26T00:00:00.000Z","type":"string"}},{"name":"event_name","required":true,"in":"query","description":"The event name of the reports","schema":{"example":"install","enum":["install","purchase"],"type":"string"}},{"name":"take","required":false,"in":"query","description":"The number of reports to get","schema":{"minimum":1,"example":10,"type":"number"}},{"name":"page","required":false,"in":"query","description":"The page number of the reports","schema":{"minimum":0,"example":1,"type":"number"}}],"responses":{"200":{"description":"Aggregated reports found","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetAggregatedReportsByAdIdResponse"}}}}},"400":{"description":"Validation error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpErrorSchema"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpErrorSchema"},"example":{"message":"Internal server error","statusCode":500}}}}},"summary":"Get aggregated reports by ad_id","tags":["Reports"]}}},"info":{"title":"Impulse Test Case API","description":"Tsarev Mykola Test Case API for Impulse","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"HttpErrorSchema":{"type":"object","properties":{"message":{"description":"Error message","oneOf":[{"example":"Validation error","type":"string"},{"type":"array","description":"As array of strings in case of validation input data","example":["field is required","field is not valid"],"items":{"type":"string"}}],"example":["field is required","field is not valid"]},"statusCode":{"type":"number","description":"Error status code","examples":[400,401,404,500],"example":400},"error":{"type":"string","examples":["ValidationError","BadRequestError"],"example":"ValidationError"}},"required":["message","statusCode"]},"Pagination":{"type":"object","properties":{"next":{"type":"string","description":"The next page url"},"totalPages":{"type":"number","description":"The total number of pages"}},"required":["next","totalPages"]},"GetAggregatedReportsByAdIdResponse":{"type":"object","properties":{"data":{"items":{"type":"array"},"type":"array"},"pagination":{"description":"The pagination","allOf":[{"$ref":"#/components/schemas/Pagination"}]}},"required":["data","pagination"]}}}}